Massively simplified getSpawn() function to remove pointless 1024x iteration
j = k = u = v = 0, so the code for swapping the signs of u and v simply swaps the sign of 0 to -0 = 0, while j += u is simply adding 0 to j. Therefore, j and k never change from their initial value of 0, meaning the if conditional always returns true--and the findServerSpawn() call is always exactly the same, making repeating it 1024 times unnecessary.