Added ocean detection for Beta 1.7
(generator.c)
- getMinCacheSize(): Allocate extra cache for Beta ocean column noise buffer

(noise.c)
- samplePerlin(): Reverted previous change, as the attempted optimization had
      unintended side-effects for 1.18+ generation and had no measurable
      performance improvement.

(layers.h)
- Renamed two OctaveNoise members of struct SurfaceNoiseBeta
- Added new struct SeaLevelColumnNoiseBeta
- Added headers for four new functions defined in layers.c

(layers.c)
- initSurfaceNoiseBeta(): Reflect changes to names in layers.h
- sampleBiomeNoiseBeta(): When nptype == 1, returned humidity value is now
      multiplied by temperature to reflect the behavior seen in biome gen
- Added new function genColumnNoise():
      Generates values used to calculate terrain noise columns and places
      them in a SeaLevelColumnNoiseBeta struct pointed to by the provided
      pointer.
- Added new function processColumnNoise():
      Generates a partial noise column for finding the block at sea level
      using data pointed to by the provided SeaLevelColumnNoiseBeta pointer
- Added new function sampleBlocks():
      Using adjacent noise columns, determine whether the blocks at y=64 are
      solid and place results in array pointed to by provided pointer
- Added new function sampleBetaBiomeOneBlock():
      Generates new noise columns for each sampled block and skips the
      diagonal traversal with cache buffer, as no noise columns data can be
      re-used when scale >= 8.
- genBetaBiomeNoiseScaled():
          Previous "no ocean" code is now locked behind (noOcean || scale >= 8)
      condition. Calls sampleBiomeNoiseBeta() when noOcean is true, and calls
      sampleBetaBiomeOneBlock() otherwise.
          Else, the generated region is traversed in 4x4 block sections
      corresponding to terrain noise columns. Columns are saved for re-use in
      future passes in order to minimize unnecessary Perlin generation, and
      region traversal is done in diagonal stripes in order to minimize extra
      cache size needed (on average) for the saved noise sample buffer.

My original plan for the diagonal traversal was to store SeaLevelColumnNoise
*pointers* in the extra cache space, which would be dynamically allocated
when generated, but since extra cache space will only exceed the original
cache space when the generated area is so small that memory concerns are
trivial, this plan was abandoned in favor of storing the structs directly in
the cache. This approach also eliminates the risk of memory leaks.

Currently, with oceans, this code takes about 20 seconds to generate the
biomes for a 4096 x 4096 block region. Since this approach to ocean-finding
deals in 4x4 noise columns, scales 2-4 are still quite slow for the same
scaled region -- about 13 seconds at 1:4 scale. At higher scales, the
diagonal approach is abandoned and four new noise columns are generated for
every midpoint block that is sampled.

I made a quick test build of Cubiomes Viewer, and while the ocean-finding
code works, the regions are very slow to show up in the view window, and
changing the zoom level can cause regions to reset when zooming in both
directions. Perhaps some optimizations can be made in Cubiomes Viewer in
terms of handling cache and displaying generated biomes at different scales,
but I'm not familiar enough with that project to suggest anything specific.
I tried to cut out as much unnecessary processing as possible in my
implementation of the inherently cpu-intensive ocean-finding algorithm, but
it's certainly possible optimizations can be made in the generation code
itself.