getSpawn Spiral Iterator Fix + Include Frozen Oceans in findServerSpawn Check
Cubitect/cubiomes-viewer#219 brought my attention to getSpawn's spiral iterator, which I checked against the actual Java code; the actual code in 1.19 only runs for 121 iterations ( = a 11x11 area), not for 1024 ( = a 32x32 area), which may explain part of (though not all of) the discrepencies. The then-fixed if condition to determine whether to subsequently call findServerSpawn would also always return true as well, so I removed it.

In addition, I added frozen oceans and deep frozen oceans as valid spawn biomes in findServerSpawn; since the game itself only checks for solid non-waterlogged surface blocks, I've found in practice that one usually ends up spawning on an ice floe/iced-over ocean as opposed to the game dropping the player on the nearest mainland, which is where the Cubiomes algorithm would set the spawnpoint otherwise.